go sql_driver: add sample sql driver
TODO: We probably don't want to use anything that looks like
this. We have discussed either generated code or an ORM.

This is an example mostly to show how the rest of code might
interact with such a thing.

We may (like this) want to set it behind a small interface for
simplicity in testing and to be replacable later.