database, handlers, middleware: Store Interfaces
Refactored database package to exposing a Store interface for data
interaction instead of exposing the DB. The primary inspiration for this
was to allow sane unit testing of handlers, but also separates concerns
of data access from the handlers.

Changed the usages of database.DB in handlers and middleware to using an
exported GORM User Store.

Fixed potential panic in Me handler, check for successful cast.