database, handlers: Setting database.Users to ref
It makes more sense for this exported member to contain a reference. We
can't technically enforce that, but hopefully future stores will follow
suit.

Refactored the mockUserStore setup in the tests to not use a global
Users table, instead making the Users table a member that is reset every
test. This is to avoid tests accidentally polluting a shared data table.