Data Layer Refactor and Handler Tests (#60)
* handlers: added ping handler unit test

* database, handlers, middleware: Store Interfaces

Refactored database package to exposing a Store interface for data
interaction instead of exposing the DB. The primary inspiration for this
was to allow sane unit testing of handlers, but also separates concerns
of data access from the handlers.

Changed the usages of database.DB in handlers and middleware to using an
exported GORM User Store.

Fixed potential panic in Me handler, check for successful cast.

* handlers: Added user handler tests

Using a mocked UserStore interface, wrote Unit Tests for the handlers in
users.go. Also ran go mod tidy.

* handlers: Return all user data from Me

Me endpoint will use the object from the database instead of the JWT to
send data back.

* handlers: Appease golangci-lint

io.NopCloser appears to also work but technically it's in the ioutil
package so whatever this is fine idk what the difference is.

* handlers: Check error of utility function

I forgot to check the error of the function that creates the request
bodies in the tests.

* database: Genericize UserUniquenesError message

This is the message that gets sent back to the client, who doesn't need
to know why user creation failed. The data is still there in the error
object itself if we want to perform different business logic on it
later.

* handlers: Better fail log and JSON assertion

We should get into a habit of using t.Fatalf instead of silently
failing, so that should start with this PR. Also included an assertion
for the JSON response.

* handlers, router: Renamed PingHandler to Ping

The other handlers were renamed to not say "Handler" so reflecting that
with Ping as well.

* database, handlers: Setting database.Users to ref

It makes more sense for this exported member to contain a reference. We
can't technically enforce that, but hopefully future stores will follow
suit.

Refactored the mockUserStore setup in the tests to not use a global
Users table, instead making the Users table a member that is reset every
test. This is to avoid tests accidentally polluting a shared data table.

* handlers: Added t.Fatalf failure messages

Added statusCodeMistmatchMessage failure message function and use it in
all tests for more descriptive failure logs.

* router: Removed named handlers import

Was probably left in by accident, but the named import is redundant
because we were naming it the exact same thing as the package.

* handlers: Added PingResponse type

We will thank ourselves later if we start typing our responses. It will
also make for more sane testing, and defines our API response contract
more explicitly.

* handlers: Added types for user handler responses

Ass with the previous commit, adding types for each handlers responses.

* handlers: Test user handler JSON responses

Added tests for user handler JSON responses using the new response
types.

* remove useless UserResponse struct

* handlers: Consolidated response structs

There were some duplicated response structs in the original set, so
consolidated them into different structs for the type of data a handler
will respond with. Moved the ErrorResponse into a new common.go file,
since that same error response will likely be reused in future handlers.

* database, handlers: Remove unnecessary User in err

The User struct doesn't need to be part of the error struct as we don't
do any logic on them yet, and usually whoever receives the error has a
reference to the user anyway.

* handlers: Testing Content-Type header

Testing the Content-Type header is the expected `application/json`,
since if the response has a JSON body it should have the expected
content type. Also added helper methods for common assertion patterns.

* database, handlers: Remove unnecessary err fields

The UserNotFoundError also had unnecessary fields ID and Email. These
are things that callers will always have when receiving the error, so
making them part of the Error struct is superfluous.

* database: explanatory comments, error name format

Added comments that explain the relationships between `users.go` and
`gorm_user_store.go` more clearly.

Fixed staticcheck error, failed to follow convention on error names in
previous commit.

* handlers: Removed tests

The tests are living on a new branch based on this one. This branch is
now for the purpose of the User Store changes.

* handlers: Use renamed ErrUserNotFound

I neglected to reflect the renamed error in handlers/user.go

* handlers: Added necessary return function

Re-added a return statement that terminates the function as intended on
success.

* database, handlers: Removed ErrorColumn field

The ErrorColumn field is not clear cut to retrieve from a Postgres Wire
error for a unique_violation, so rather than further complicate this PR
I simplified the UserUniquenessError to become ErrUserNotUnique.

* handlers: ErrorResponse returns an array of error

The common ErrorResponse now responds with an array of errors in the
event that multiple errors occurred during the request.

* handlers: Added SuccessResponse with top level key

Using `data` as a top level key for successful responses.

Co-authored-by: Ted Wollman <25165500+TheTedder@users.noreply.github.com>